+++ src/components/Icon/Icon.tsx
@@ -6,7 +6,7 @@ import { IconType } from "./IconConstants";
 import CustomSvgIcon from "./CustomSvgIcon/CustomSvgIcon";
 import FontIcon from "./FontIcon/FontIcon";
 import useIconProps from "./hooks/useIconProps";
-import { VibeComponentProps, VibeComponent, MouseEventCallBack, SubIcon } from "../../types";
+import { VibeComponentProps, VibeComponent, MouseEventCallBack, SubIcon, withStaticProps } from "../../types";
 
 // eslint-disable-next-line no-unused-vars,@typescript-eslint/no-unused-vars,@typescript-eslint/no-empty-function
 const CLICK_NOOP = (_event: React.MouseEvent) => {};
@@ -156,8 +156,6 @@ const Icon: VibeComponent<IconProps, HTMLElement> & { type?: typeof IconType } =
   }
 );
 
-Object.assign(Icon, {
+export default withStaticProps(Icon, {
   type: IconType
 });
-
-export default Icon;