+++ src/components/Tabs/TabPanels/TabPanels.tsx
@@ -7,6 +7,7 @@ import { TabPanelsAnimationDirection } from "./TabPanelsConstants";
 import { TabPanelProps } from "../TabPanel/TabPanel";
 import { ComponentDefaultTestId, getTestId } from "../../../tests/test-ids-utils";
 import { getStyle } from "../../../helpers/typesciptCssModulesHelper";
+import { withStaticProps } from "../../../types";
 import styles from "./TabPanels.module.scss";
 
 export interface TabPanelsProps extends VibeComponentProps {
@@ -67,8 +68,9 @@ const TabPanels: FC<TabPanelsProps> & {
 );
 
 Object.assign(TabPanels, {
-  isTabPanels: true,
-  animationDirections: TabPanelsAnimationDirection
+  isTabPanels: true
 });
 
-export default TabPanels;
+export default withStaticProps(TabPanels, {
+  animationDirections: TabPanelsAnimationDirection
+});