+++ src/tests/constants.ts
@@ -1,6 +1,7 @@
 export enum ComponentDefaultTestId {
   // Don't remove next line
   // plop_marker:default-data-testid-declarations
+  TEXT = "text",
   COMBOBOX = "combobox",
   COLOR_PICKER = "color-picker",
   CHECKBOX = "checkbox",